Fixed audio switching among different number of channels.

+++ mrViewer/src/video/mrvGLShape.h
@@ -52,7 +52,7 @@ class GLShape
 	a = ai;
      }
      float r, g, b, a;
-     double pen_size;
+     float pen_size;
 };
 
 class GLPathShape : public GLShape
